fix: temp move registrar managed name to plugin-local
