+++ apps/ensindexer/src/plugins/basenames/basenames.plugin.ts
@@ -53,8 +53,6 @@ export const config = createConfig({
 
 export const activate = activateHandlers({
   pluginName,
-  // the shared Registrar handler in this plugin indexes direct subnames of '.base.eth'
-  registrarManagedName: "base.eth",
   namespace,
   handlers: [
     import("./handlers/Registry"),
