+++ apps/ensindexer/src/plugins/basenames/handlers/Registrar.ts
@@ -20,7 +20,6 @@ const tokenIdToLabelHash = (tokenId: bigint): LabelHash => uint256ToHex32(tokenI
 
 export default function ({
   pluginName,
-  registrarManagedName,
   namespace,
 }: ENSIndexerPluginHandlerArgs<PluginName.Basenames>) {
   const {
@@ -31,10 +30,11 @@ export default function ({
     handleNameTransferred,
   } = makeRegistrarHandlers({
     pluginName,
-    registrarManagedName,
+    // the shared Registrar handlers in this plugin index direct subnames of '.base.eth'
+    registrarManagedName: "base.eth",
   });
 
-  const registrarManagedNode = namehash(registrarManagedName);
+  const registrarManagedNode = namehash("base.eth");
 
   // support NameRegisteredWithRecord for BaseRegistrar as it used by Base's RegistrarControllers
   ponder.on(namespace("BaseRegistrar:NameRegisteredWithRecord"), async ({ context, event }) => {
