+++ apps/ensindexer/src/plugins/lineanames/handlers/Registrar.ts
@@ -20,7 +20,6 @@ const tokenIdToLabelHash = (tokenId: bigint): LabelHash => uint256ToHex32(tokenI
 
 export default function ({
   pluginName,
-  registrarManagedName,
   namespace,
 }: ENSIndexerPluginHandlerArgs<PluginName.Lineanames>) {
   const {
@@ -31,10 +30,11 @@ export default function ({
     handleNameTransferred,
   } = makeRegistrarHandlers({
     pluginName,
-    registrarManagedName,
+    // the shared Registrar handlers in this plugin index direct subnames of '.linea.eth'
+    registrarManagedName: "linea.eth",
   });
 
-  const registrarManagedNode = namehash(registrarManagedName);
+  const registrarManagedNode = namehash("linea.eth");
 
   ponder.on(namespace("BaseRegistrar:NameRegistered"), async ({ context, event }) => {
     await handleNameRegistered({
